<!--
Thanks for contributing a pull request! Please ensure you have taken a look at
the contribution guidelines: https://github.com/scikit-learn/scikit-learn/blob/main/CONTRIBUTING.md
-->

#### What does this implement/fix? Explain your changes.
Currently, when specifying `n_jobs` for `plot_partial_dependence`, it is parallelized by concurrently computing the partial dependences *per feature* (by calling the function `partial_dependence`). When only one feature is to be plotted, specifying `n_jobs` will lead to no performance boost at best. (Or more general, when `n_jobs` > number of features).

Apart from the fact that I don't think this is documented anywhere, this can be implemented differently to also provide a speedup when the number of features is smaller than n_jobs.

I moved the parallelization to partial_dependence instead. It divides up the grid for which the values are computed and parallelizes among those different parts of the grid. This has two benefits:

- Specifying `n_jobs` now also leads to a performance boost when `n_jobs` > number of features
- `partial_dependence` itself now has a parameter `n_jobs` and can therefore also be parallelized


#### Timing results

Some results of timeit before and after the changes (on my 4-core MacBook, borrowed some code from the user guide):

```
import pandas as pd
from sklearn.datasets import fetch_california_housing
from sklearn.neural_network import MLPRegressor
from sklearn.inspection import plot_partial_dependence

cal_housing = fetch_california_housing()
X = pd.DataFrame(cal_housing.data, columns=cal_housing.feature_names)
y = cal_housing.target

est = MLPRegressor(hidden_layer_sizes=(64, 64),
                   learning_rate_init=0.01,
                   early_stopping=True)

est.fit(X, y)

def generate_plots(features, n_jobs):
    return plot_partial_dependence(
        est, X, features, kind="both", subsample=50,
        n_jobs=n_jobs, grid_resolution=100, random_state=0)
```

One feature, n_jobs=1:
```
%%timeit

features = ['MedInc']
generate_plots(features, n_jobs=1)
```
Before change: `2.66 s ± 35.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`
After change: `2.71 s ± 29.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`

One feature, n_jobs=4:
```
%%timeit

features = ['MedInc']
generate_plots(features, n_jobs=4)
```
Before change: `3.47 s ± 596 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`
After change: `1.88 s ± 40 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)`