@timo-stoettner 2021-02-13
Thanks @jjerphan! I followed your suggestion and ran some more extensive benchmarks. In some cases that I haven't tested before `main` indeed appears to be faster. I'll have to figure out why and adapt the code accordingly. I changed the PRs title back to WIP for now.