@NicolasHug 2021-02-17
Thanks for the proposal @timo-stoettner 

> In some cases that I haven't tested before main indeed appears to be faster. I'll have to figure out why and adapt the code accordingly.

as far as I understand this PR splits the data into `n_jobs` chunks so each job works on about `grid_resolution // effective_n_jobs` data points. When there aren't too many data points and when the computation is pretty fast (e.g. in the recursion method),  spawning the jobs can induce a significant overhead. This is even more true when there are many plots to make, as this will spawn `effective_n_jobs * n_features` jobs.

Parallelizing over samples might be a good idea but indeed we need to figure out in which cases this makes sense. Typically if `effective_n_job <= n_features` I think we should stick to the current parallelization

> I would suggest performing a benchmark between main and your branch on a varying number of features and jobs

and `grid_resolution` ;)