@timo-stoettner 2021-02-20
@NicolasHug As suggested I've added different parallelization strategies based on the comparison of `effective_n_jobs` and `n_features`. Also I've added results of some more extensive benchmarks to the PR description.